requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# ML/AI dependencies
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
torch==2.1.0
faiss-cpu==1.7.4

# LLM dependencies
openai==1.3.5
langchain==0.0.350
tiktoken==0.5.1

# Data processing
pandas==2.1.3
pyyaml==6.0.1

# API and async
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
memory-profiler==0.61.0
matplotlib==3.8.2